requirements.txt
python-doctr
rapidfuzz==2.15.1
torch
torchvision
urllib3==1.26.6
weasyprint